1384 Moves WfoAdvancedTable to its own folder
