+++ packages/orchestrator-ui-components/src/components/WfoTable/WfoAdvancedTable/index.ts
@@ -1,0 +1,1 @@
+export * from './WfoAdvancedTable';
