+++ packages/orchestrator-ui-components/src/utils/index.ts
@@ -1,6 +1,8 @@
 export * from './csvDownload';
 export * from './date';
 export * from './environmentVariables';
+export * from './filterData';
+export * from './getDefaultTableConfig';
 export * from './getEnvironmentVariables';
 export * from './getObjectKeys';
 export * from './getProductNamesFromProcess';
@@ -8,9 +10,8 @@ export * from './getQueryVariablesForExport';
 export * from './getStatusBadgeColor';
 export * from './getTypedFieldFromObject';
 export * from './onlyUnique';
-export * from './uuid';
+export * from './optionalArray';
 export * from './resultFlattener';
 export * from './strings';
-export * from './optionalArray';
-export * from './filterData';
 export * from './toObjectWithSortedKeys';
+export * from './uuid';
